+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Last Changes
 
+- [#6](https://github.com/LaxarJS/grunt-laxar/issues/6): css_merger: fixed imports by URL
 - [#36](https://github.com/LaxarJS/grunt-laxar/issues/36): css_merger fix incorrect CSS output path and `url()`s for windows
 - [#35](https://github.com/LaxarJS/grunt-laxar/issues/35): widget dependencies: fixed URL generation for windows
+
 
 ## v0.7.0